// src/app/Models/Pagos/TablaAmortizacionRequest.ts
export class TablaAmortizacionRequest {
  idDistribuidora: number;
  folioVale: string
  montoVenta: number;
  nombre: string;
  aPaterno:string;
  aMaterno: string;
  fechaNacimiento: string;
  calle: string;
  numExt: string;
  colonia: string;
  cP: string;
  municipio: string;
  estado: string;
  ine: string;
  sexo: string;
}